fix(e2e): olm test reorganization
