+++ .github/workflows/install.yml
@@ -70,7 +70,7 @@ concurrency:
   cancel-in-progress: true
 
 jobs:
-  install-it:
+  non-olm:
     runs-on: ubuntu-20.04
     steps:
     - name: Checkout code
@@ -87,13 +87,13 @@ jobs:
           -q "${{ github.event.inputs.log-level }}" \
           -t "${{ github.event.inputs.test-filters }}"
 
-    - name: Install Tests
+    - name: Install and upgrade test (OLM)
       uses: ./.github/actions/e2e-install
       with:
         cluster-config-data: ${{ secrets.E2E_CLUSTER_CONFIG }}
         cluster-kube-config-data: ${{ secrets.E2E_KUBE_CONFIG }}
 
-  upgrade-it:
+  olm:
     runs-on: ubuntu-20.04
 
     steps:
@@ -111,7 +111,7 @@ jobs:
           -q "${{ github.event.inputs.log-level }}" \
           -t "${{ github.event.inputs.test-filters }}"
 
-    - name: Upgrade Tests
+    - name: Install and upgrade test (OLM)
       uses: ./.github/actions/e2e-upgrade
       with:
         cluster-config-data: ${{ secrets.E2E_CLUSTER_CONFIG }}
