+++ e2e/install/cli/upgrade_test.go
@@ -20,7 +20,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package upgrade
+package cli
 
 import (
 	"os"
